feat: 수동 번호 입력 기능 구현
