+++ src/main/kotlin/lotto/view/input/InputView.kt
@@ -8,6 +8,7 @@ import lotto.domain.LottoPurchaseCount
 interface InputView {
     fun getPurchaseAmount(): LottoPurchaseAmount
     fun getManualPurchaseCount(): LottoPurchaseCount
+    fun getManualNumbers(manualPurchaseCount: LottoPurchaseCount): List<LottoNumberPackage>
     fun getWinningNumbers(): LottoNumberPackage
     fun getBonusNumber(winningNumbers: LottoNumberPackage): LottoNumber
 }
