+++ src/main/kotlin/lotto/view/result/ResultView.kt
@@ -5,7 +5,7 @@ import lotto.domain.LottoResult
 import lotto.domain.LottoTicket
 
 interface ResultView {
-    fun showPurchaseCount(purchaseCount: LottoPurchaseCount)
+    fun showPurchaseCount(manualPurchaseCount: LottoPurchaseCount, automaticPurchaseCount: LottoPurchaseCount)
     fun showLottoTicketNumber(lottoTicket: LottoTicket)
     fun showResultStatistics(result: LottoResult)
 }
