+++ src/test/kotlin/lotto/domain/LottoNumberPackageTest.kt
@@ -5,7 +5,9 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
+import org.junit.jupiter.params.provider.EmptySource
 import org.junit.jupiter.params.provider.MethodSource
+import org.junit.jupiter.params.provider.ValueSource
 import java.util.stream.Stream
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
@@ -212,5 +214,59 @@ internal class LottoNumberPackageTest {
             Arguments.of(listOf(1, 2, 3, 4, 5, 6), listOf(1, 2, 3, 4, 5, 6), "13"),
             Arguments.of(listOf(7, 8, 9, 10, 11, 12), listOf(7, 8, 9, 10, 11, 12), "13"),
         )
+    }
+
+    @ParameterizedTest
+    @EmptySource
+    fun `당첨 번호에 빈 문자열을 입력하면 IllegalArgumentException이 발생한다`(input: String) {
+        Assertions.assertThatThrownBy {
+            LottoNumberPackage.from(input)
+        }.isInstanceOf(IllegalArgumentException::class.java)
+    }
+
+    @ParameterizedTest
+    @ValueSource(
+        strings = [
+            "aaa, 1, 2, 3, 4, 5",
+            "테스트, 1, 2, 3, 4, 5",
+            "!, 1, 2, 3, 4, 5",
+            "/, 1, 2, 3, 4, 5"
+        ]
+    )
+    fun `당첨 번호에 숫자가 아닌 값을 입력하면 IllegalArgumentException이 발생한다`(input: String) {
+        Assertions.assertThatThrownBy {
+            LottoNumberPackage.from(input)
+        }.isInstanceOf(IllegalArgumentException::class.java)
+    }
+
+    @ParameterizedTest
+    @ValueSource(
+        strings = [
+            "1",
+            "1, 2",
+            "1, 2, 3",
+            "1, 2, 3, 4",
+            "1, 2, 3, 4, 5",
+            "1, 2, 3, 4, 5, 6, 7",
+        ]
+    )
+    fun `당첨 번호가 6개가 아니면 IllegalArgumentException이 발생한다`(input: String) {
+        Assertions.assertThatThrownBy {
+            LottoNumberPackage.from(input)
+        }.isInstanceOf(IllegalArgumentException::class.java)
+    }
+
+    @ParameterizedTest
+    @ValueSource(
+        strings = [
+            "1, 2, 3, 4, 5, 6",
+            "21, 32, 43, 4, 15, 27"
+        ]
+    )
+    fun `당첨 번호가 6개를 입력하면 정상적으로 WinningInfo 가 생성된다`(input: String) {
+        val winningInfo = LottoNumberPackage.from(input)
+
+        assertThat(winningInfo).isNotNull
+        assertThat(winningInfo.size()).isEqualTo(LottoNumberPackage.LOTTO_GAME_NUMBER_COUNT)
     }
 }
